% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_globals.R
\name{attach_globals}
\alias{attach_globals}
\title{attach_globals}
\usage{
attach_globals(..., .data, .f, .env = parent.frame())
}
\arguments{
\item{...}{a list of functional arguments}

\item{.data}{an optional object to map from}

\item{.f}{an optional transform function to apply to mapping}

\item{.env}{a target export environment, DEFAULT parent.frame()}
}
\description{
Export arguments to global namespace
}
