% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{calculate_panel_statistics}
\alias{calculate_panel_statistics}
\title{Calculate statistics for a given panel}
\usage{
calculate_panel_statistics(.x, panel)
}
\arguments{
\item{panel}{A assay panel, or a list with (name,markers)}
}
\description{
Calculate statistics for a given panel
}
