% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty.R
\name{check_empty}
\alias{check_empty}
\title{check_empty}
\usage{
check_empty(data)
}
\arguments{
\item{data}{an object that you want to check for "emptiness"}
}
\value{
logical TRUE/FALSE
}
\description{
checks an object to determine if it is "empty".
}
