% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_csv.R
\name{collate_csv}
\alias{collate_csv}
\title{collate_csv}
\usage{
collate_csv(directory, group, response = "Value")
}
\arguments{
\item{directory}{File path specifying the directory containing all the csv files to be collated.}

\item{group}{Character string specifying group variable name. Ex: "Ethnicity"}

\item{response}{Character string specifying response variable name. Defaults to "Value." Must refer to a consistent name across all pivot tables to be collated}
}
\description{
Collates dataframes of the form created by surveyPivot across multiple years
}
\details{
Collates csv files from multiple years containing pivot tables created with surveyPivot. Variable names must be consistent across all files in the directory. The csv file names must contain their year (ex: "survey_2020.csv").
}
\keyword{collate}
\keyword{read_csv}
\keyword{time-series}
