% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deprecated}
\alias{\%:>\%}
\alias{\%|>\%}
\alias{deprecated}
\title{Pipe an object forward to expression (deprecated)}
\usage{
. \%:>\% expr

. \%|>\% expr
}
\arguments{
\item{.}{object}

\item{expr}{expression}
}
\description{
Pipe an object forward to expression (deprecated)
}

