% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engagePlot.R
\name{engagePlot}
\alias{engagePlot}
\title{engagePlot}
\usage{
engagePlot(data, major, etype, save = FALSE, save_dir = NULL)
}
\arguments{
\item{data}{Dataframe of with class year, major, and individual student engagement tallies for events (e_attend), coach (coach_appt), peer advisor (pa_appt), and coach or peer advisor (c_pa_appt)}

\item{major}{Major of interest (character). Must match major spelling and case exactly as found in data}

\item{etype}{Engagement type (character). "Coach or Peer Advisor", "Coach", "Peer Advisor", or "Events"}

\item{save}{Do you wish to save the plot output to the session directory? Default is FALSE}

\item{save_dir}{If save==TRUE, specify what directory (or file path to that directory, omitting the top-level working directory) in which you wish to save the file}
}
\description{
Engagement Plot Function
}
\details{
Plots type-specific engagement over all four class years by major.
}
\examples{
engagePlot(engage_df, "Mathematics", "Coach", save=TRUE, save_dir="piper")
engagePlot(engage_df, "Mathematics", "Coach or Peer Advisor", save=TRUE, save_dir = file.path("piper","engagement"))
}
\keyword{engagement}
