% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{error_guard}
\alias{error_guard}
\title{error_guard}
\usage{
error_guard(expr, args, env, stack, pipe)
}
\arguments{
\item{expr}{a valid R expression to error check}

\item{args}{optional meta information}

\item{env}{parent environment pointer}

\item{stack}{pipe dependency stack}

\item{pipe}{piper pipe pointer}
}
\description{
Custom internal error handler.
}
