% Generated by roxygen2 (4.0.1): do not edit by hand
\name{first-argument piping}
\alias{\%>>\%}
\alias{first-argument piping}
\title{Pipe an object forward as the first argument to a function}
\usage{
. \%>>\% fun
}
\arguments{
\item{.}{The object to be piped as the first argument}

\item{fun}{The function call to evaluate with the piped object as the first argument.}
}
\description{
The \code{\%>>\%} operator evaluates the function call on the right-hand side
with the left-hand side object being the first argument.
}
\examples{
\dontrun{
rnorm(100) \%>>\% plot

rnorm(100) \%>>\% plot(col="red")

rnorm(1000) \%>>\% sample(size=100,replace=F) \%>>\% hist
}
}

