% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fun}
\alias{fun}
\title{Evaluate a lambda expression for Pipe object}
\usage{
fun(value, expr, envir = parent.frame(2L))
}
\arguments{
\item{value}{value}

\item{expr}{the lambda expression in the following forms:

1. An expression with \code{.} representing \code{value}

2. \code{x -> f(x)}

3. \code{x ~ f(x)}}

\item{envir}{the environment to evaluate \code{expr}.}
}
\description{
Evaluate a lambda expression for Pipe object
}

