% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{\%-\%}
\alias{\%-\%}
\title{Fail-safe object initializer & validator}
\usage{
lhs \%-\% rhs
}
\arguments{
\item{lhs}{a valid R object reference}

\item{rhs}{a valid R object reference}
}
\description{
Fail-safe object initializer & validator
}
