% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSheet.R
\name{importSheet}
\alias{importSheet}
\title{importSheet}
\usage{
importSheet(link, email, sheet, range = NULL, col_names = TRUE, ...)
}
\arguments{
\item{link}{Link to Google Sheets document containing the data. May be a full url or a Google Sheets ID (the string between "../d/" and "/edit..." in the url).}

\item{email}{Email address which has access to the Google Sheets document.}

\item{sheet}{Character string specifying the name of the spreadsheet within the Google Sheets document which contains the desired data.}

\item{range}{Optional. Character string specifying the range to be imported (e.g., 'A1:G100').}

\item{col_names}{Defaults to TRUE, using the first row in the spreadsheet as the column names. A character vector of custom column names may be passed to this argument.}

\item{...}{Other read_sheet arguments. See ?read_sheet for more.}
}
\description{
Imports data from a specified Google Sheets spreadsheet
}
\details{
Wraps several googlesheets4 functions to import a specified sheet (and range if desired) from a Google Sheets document. Aside from link and email, all googlesheets4::read_sheet() arguments may be passed to importSheet()
}
\keyword{googlesheets4}
\keyword{import}
