% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{impute_missing_with_values}
\alias{impute_missing_with_values}
\title{Impute missing values with specified alternate values}
\usage{
impute_missing_with_values(x, values, offset = 0)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} to impute missing values from.}

\item{values}{A vector of values (recycle rules per \code{\link[dplyr:coalesce]{dplyr::coalesce()}}) to use as imputed values.}

\item{offset}{A value to add to values (default 0).}
}
\value{
A matrix the same dimensions as the default \code{assay} slot
changed such that missing values are imputed with the specified alternate values.
}
\description{
Impute missing values with specified alternate values
}
