% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{is_comparable}
\alias{is_comparable}
\title{Test if sample and reference are comparable}
\usage{
is_comparable(sample, reference)
}
\arguments{
\item{reference}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} representing the reference population.}
}
\value{
Logical value indicating if sample and reference are comparable.
}
\description{
Test if sample and reference are comparable
}
