% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{is_marker_in_panels}
\alias{is_marker_in_panels}
\title{Determine if marker is in a list of panels}
\usage{
is_marker_in_panels(marker, panels)
}
\arguments{
\item{marker}{A string representing a marker}

\item{panels}{A list of panels, each of which has a markers component.}
}
\value{
A list of named logicals, with the name representing the
panel and the value representing if the marker is within the panel.
}
\description{
A list of panels each contains a set of markers. Test
if a given marker is in each of the panels. Returns a (named) logical vector
indicating if the marker is the panel.
}
\examples{
\dontrun{
markers_in_panels("CDN2A_ALLEAGALPNAPNSYGR", panels)
}
}
