% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_format}
\alias{label_format}
\title{Format labels with padding on all sides}
\usage{
label_format(s)
}
\arguments{
\item{s}{A vector of strings}
}
\value{
A vector of strings consisting of padded, line-separated components
of the input assay fields.
}
\description{
The assay label is a composite field. This function splits the
composite label into components and pads them to a max width. Then combines
them into a single string.
}
\details{
This function is a wrapper function for ease of use. It uses
\code{\link[=split_assay_id]{split_assay_id()}} to split a composite label into
components. Then it uses \code{\link[=label_pad]{label_pad()}} to pad and concatenate the results into
single strings.
}
\examples{
\dontrun{
label_format("K2C5_LAELEEALQK")
[1] "                      \n         K2C5         \n      LAELEEALQK      \n                      "
}
}
