% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_pad}
\alias{label_pad}
\title{Pad labels to max width}
\usage{
label_pad(..., extra_width = 12)
}
\arguments{
\item{...}{A series of labels (strings) to concatenate with the same width}

\item{extra_width}{Additional width to add to longest string component (default = 6)}
}
\value{
A string with extra padding and concatenation
}
\description{
Given input parameters (labels), find the largest label and pad
all strings to that width. Then concatenate into a single string.
}
\details{
When plotting figures in \code{\link[plotly:plot_ly]{plotly::plot_ly()}} the padding
is not always great. This function assumes the goal is to create
a multi-line string consisting of elements of the input. It determines
the maximum width of input labels, then centers all labels to that width
using spaces. The argument \code{extra_width} can be used to increase the
padding beyond the widest string.

The padded strings are then concatenated together into a single
string, with a newline separating the strings. Additionally, an
extra line is added before and after the strings.
}
\examples{
\dontrun{
label_pad("Line1","This is Line2", extra_width=0)

[1] "             \n    Line1    \nThis is Line2\n             "
}
}
