% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{module_.push}
\alias{module_.push}
\title{module_.push}
\usage{
module_.push(
  .this = {
 },
  .with = {
 },
  ...
)
}
\arguments{
\item{.this}{state initializers}

\item{.with}{expression constructors}

\item{...}{a list of functional arguments}
}
\description{
Populate piper module stack.
}
