% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{old_percent}
\alias{old_percent}
\title{Calculate percent of values for samples}
\usage{
old_percent(x, values, use_na = TRUE)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} representing samples.}

\item{values}{Values to use as the maximum/total from which to calculate percentages.}

\item{use_na}{If TRUE (default), keep NA values from sample as NA percent. Otherwise, use 0 percent.}
}
\value{
A fractional percentage calculated against \code{values} for the default \code{assay} slot
of \code{x}.
}
\description{
Calculate percent of values for samples
}
\examples{
\dontrun{
irisse <- SummarizedExperiment::SummarizedExperiment(t(iris[,1:4]))
context_percent_max(irisse[,5], irisse)
}
}
