% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgplot.R
\name{pgplot}
\alias{pgplot}
\title{pgplot}
\usage{
pgplot(
  data,
  x,
  y,
  group,
  facet,
  year = "year",
  title = "DEFAULT TITLE",
  subtitle = "DEFAULT subtitle",
  ylab = "DEFAULT ylab",
  xlab = "DEFAULT xlab",
  caption = NULL,
  palette = "Dark2"
)
}
\arguments{
\item{data}{Dataframe.}

\item{x}{Variable to be mapped on the x-axis}

\item{y}{Variable to be mapped on the y-axis}

\item{group}{Group variable (e.g., ethnicity), which will be delineated by color}

\item{facet}{Variable on which the plot panels (facets) will be based. Survey response is a sensible choice.}

\item{year}{Variable name denoting year. Defaults to "year".}

\item{title}{Plot title, as character string}

\item{subtitle}{Plot subtitle, as character string}

\item{ylab}{Label for y-axis, as character string}

\item{xlab}{Label for x-axis, as character string}

\item{caption}{Optional. Caption will be placed below plot.}

\item{palette}{palette from RColorBrewer. See \href{https://www.r-graph-gallery.com/38-rcolorbrewers-palettes.html}{here} for all palette options. Defaults to "Dark2".}
}
\description{
Panel plot for surveys
}
\details{
Creates multi-panel plot for survey response visualizations. Performs best when survey response strings are short and there are no more than 6 unique response values. (Name meaning: p for panel, g for gg, as in ggplot.)
}
