% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pipeR-package}
\alias{pipeR-package}
\title{The pipeR package}
\description{
The pipeR package
}
\details{
pipeR provides Pipe operator and function based on syntax which support to pipe
value to first-argument of a function, to dot in expression, by formula as lambda
expression, for side-effect, and with assignment. The set of syntax is designed to
make the pipeline highly readable.

pipeR Tutorial (\url{http://renkun.me/pipeR-tutorial}) is a highly recommended
complete guide to pipeR.
}

