% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper.brew}
\alias{piper.brew}
\title{piper.brew}
\usage{
piper.brew(pipe, library, method, version, header = "main", parent = "R++")
}
\arguments{
\item{pipe}{module root pipe, DEFAULT: modules}

\item{library}{module library}

\item{method}{a method name}

\item{version}{library version}

\item{parent}{the desired parent directory to house module assets}
}
\description{
Create new library method.
}
