% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper.load}
\alias{piper.load}
\title{piper.load}
\usage{
piper.load(module, from, .env = rlang::caller_env(), ...)
}
\arguments{
\item{module}{the name of the source asset}

\item{from}{target directory}

\item{.env}{a target environmehnt hook, DEFAULT: parent.frame()}
}
\description{
Load a newly initialized piper instance.
}
