% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper.make}
\alias{piper.make}
\title{piper.make}
\usage{
piper.make(pipe, root = "R++", mode = "0755")
}
\arguments{
\item{pipe}{the desired sub-directory containing pipe modules}

\item{root}{the desired parent directory to house module assets}

\item{mode}{rw+ permission settings}
}
\description{
Create new piper root
}
