% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper.module}
\alias{piper.module}
\title{piper.module}
\usage{
piper.module(name, pipe, parent = "R++", mode = "0755")
}
\arguments{
\item{name}{a library name}

\item{pipe}{the root pipe to use}

\item{parent}{the desired parent directory to house module assets}

\item{mode}{rw+ permission settings}
}
\description{
Create new module library
}
