% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piper.R
\name{piper.new}
\alias{piper.new}
\title{piper.new}
\usage{
piper.new(.env = parent.frame(), auto_purge = TRUE, ...)
}
\arguments{
\item{.env}{a target environmehnt hook, DEFAULT: parent.frame()}

\item{...}{a list of functional arguments}
}
\description{
Initialize a new piper instance.
}
