% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Create a lollipop plot of markers vs reference}
\usage{
plot_lollipop(sample, reference, panel)
}
\arguments{
\item{sample}{A sample SummarizedExperiment}

\item{reference}{A reference population}

\item{panel}{A panel (list with name and markers)}
}
\description{
Create a lollipop plot of markers vs reference
}
