% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Title}
\usage{
plot_radar(
  .x,
  pnum = 1,
  panel = "Tissue QC",
  use_percent = FALSE,
  peptide_list = NULL,
  axis.label.size = 5
)
}
\arguments{
\item{peptide_list}{}
}
\description{
Title
}
