% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{plot_radar_DEPRECATED}
\alias{plot_radar_DEPRECATED}
\title{Title}
\usage{
plot_radar_DEPRECATED(
  .x,
  pnum = 1,
  panel = "Tissue QC",
  use_percent = FALSE,
  peptide_list = NULL,
  axis.label.size = 4
)
}
\arguments{
\item{peptide_list}{}
}
\description{
Title
}
