% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rug.R
\name{plot_rug}
\alias{plot_rug}
\title{Generate a rug plot from proteomics panel}
\usage{
plot_rug(sample, reference, panel = NULL)
}
\arguments{
\item{sample}{A SummarizedExperiment representing a single sample}

\item{reference}{A SummarizedExperiment representing the reference population}

\item{panel}{The panel to visualize (list consisting of name and markers).}
}
\value{
A \code{\link[patchwork:patchwork-package]{patchwork::patchwork()}} panel of graphs representing a rug plot.
}
\description{
Provide slider-style readout of proteomics panel
components.
}
\examples{
\dontrun{
plot_rug(lscc[,1], lscc , panel = panels$ADC)
}
}
