% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spider.R
\name{plot_spider}
\alias{plot_spider}
\title{Title}
\usage{
plot_spider(
  .x,
  pnum = 1,
  panel = "Tissue QC",
  use_percent = FALSE,
  peptide_list = NULL
)
}
\arguments{
\item{peptide_list}{}
}
\description{
Title
}
