% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{pretty_print_table}
\alias{pretty_print_table}
\title{pretty_print_table}
\usage{
pretty_print_table(data)
}
\arguments{
\item{data}{a valid R object, e.g. data.frame}
}
\description{
Pretty print a data frame in table format with aligned columns and color palette
}
