% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{row_ecdf}
\alias{row_ecdf}
\title{Calculate row-wise empirical cumulative distribution functions}
\usage{
row_ecdf(x)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} to calculate ecdfs from.}
}
\value{
A list of ecdfs
}
\description{
Calculate row-wise empirical cumulative distribution functions
}
