% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{row_max}
\alias{row_max}
\title{Calculate row-wise maximums}
\usage{
row_max(x, na.rm = TRUE, maxval = 0)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} to calculate row maxmimums}

\item{na.rm}{Logical (default TRUE). Should NA values be removed from consideration (TRUE) or
treated as maxval (FALSE).}

\item{maxval}{A maximum value to use if there are NA's. Default is 0.}
}
\value{
A named vector with each element the maximum of the corresponding \code{x} row. Each
element is named with the corresponding row name of \code{x}.
}
\description{
Calculate row-wise maximums
}
