% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{row_min}
\alias{row_min}
\title{Calculate row-wise minimums}
\usage{
row_min(x, na.rm = TRUE)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} to calculate row minimums}

\item{na.rm}{Logical (default TRUE). Should NA values be removed from consideration (TRUE) or
treated as a 0 (FALSE).}
}
\value{
A named vector with each element the minimum of the corresponding \code{x} row. Each
element is named with the corresponding row name of \code{x}.
}
\description{
Calculate row-wise minimums
}
