% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_assay_id}
\alias{split_assay_id}
\title{Split assay ID into protein and peptide}
\usage{
split_assay_id(s)
}
\arguments{
\item{s}{A vector of assay id's.}
}
\value{
A data frame of \code{Protein} and \code{Peptide} columns corresponding
to the content of the input vector.
}
\description{
The assay id is a protein and peptide sequence combined. This
function splits them into a data frame of the two components.
}
\examples{
\dontrun{
split_assay_id("K2C7_SAYGGPVGAGIR")
}
}
