% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{summary_statistics_table}
\alias{summary_statistics_table}
\title{Title}
\usage{
summary_statistics_table(sample, reference, panels)
}
\arguments{
\item{sample}{A sample to calculate relative to}

\item{reference}{A reference population}

\item{panels}{A list of panels to consider when calculating}
}
\value{
A table of summary statistics by assay
}
\description{
Title
}
