% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_piper.R
\name{theme_piper}
\alias{theme_piper}
\title{theme_piper}
\usage{
theme_piper()
}
\description{
Custom Piper Center ggplot theme
}
\details{
Custom Piper Center visualization theme to add to ggplot objects. Built into pgplot (in-house panel plot format for Piper survey responses).
}
