% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deprecated}
\alias{\%:>\%}
\alias{\%|>\%}
\alias{deprecated}
\title{[Deprecated] Pipe an object forward to expression}
\usage{
. \%:>\% expr

. \%|>\% expr
}
\arguments{
\item{.}{object}

\item{expr}{expression}
}
\description{
[Deprecated] Pipe an object forward to expression
}

