% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{pipeR-package}
\alias{pipeR-package}
\title{The pipeR package}
\description{
The pipeR package
}
\details{
pipeR provides various styles of function chaining methods: Pipe operator,
Pipe object, and pipeline function, each representing a distinct pipeline
model yet sharing almost a common set of features: A value can be piped to
the first unnamed argument of a function, to dot symbol in an enclosed expression,
by formula as lambda expression, for side-effect, and with assignment.
The set of syntax is designed to make the pipeline more readable and fluent to
a wide variety of operations.

pipeR Tutorial (\url{http://renkun.me/pipeR-tutorial}) is a highly recommended
complete guide to pipeR.
}

