\name{pipeR-package}
\alias{pipeR-package}
\alias{pipeR}
\docType{package}
\title{
Pipeline operators for R
}
\description{
Provides various operators for chaining commands with two forward-piping mechanisms: first-argument piping and free piping.
}
\details{
\tabular{ll}{
Package: \tab pipeR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-07\cr
License: \tab MIT\cr
}



}
\author{
Kun Ren <renkun@outlook.com>

Maintainer: Kun Ren <renkun@outlook.com>
}

\keyword{ pipeline }
\seealso{

}
\examples{

}
