% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot assay results using radar plot}
\usage{
plot_radar(
  sample,
  reference,
  panel = NULL,
  axis.label.size = 4,
  show_title = FALSE
)
}
\arguments{
\item{sample}{A SummarizedExperiment object for the sample}

\item{reference}{A SummarizedExperiment object representing the reference population.}

\item{panel}{A panel list (markers, name, category)}
}
\value{
A \code{\link[plotly:plot_ly]{plotly::plot_ly()}} radar plot representing the specified
panel for a target sample against a reference population.
}
\description{
Plot an assay panel result for a sample against a reference
set using a radar plot.
}
\examples{
\dontrun{
plot_radar(lscc[,1],lscc,
  panel = list(name="ADC",markers=list("PSCA_AVG","TACD2_AAG"),
  category = "Cell Surface, RTKs & ADC Targets"))
}
}
