% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rug.R
\name{plot_rug}
\alias{plot_rug}
\title{Generate a rug plot from proteomics panel}
\usage{
plot_rug(.x, pnum = 1, panel = "Tissue QC")
}
\arguments{
\item{.x}{The \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} to visualization}

\item{pnum}{For multi-patient datasets, the patient to visualize
(should be in \code{colnames(.x)} or a numeric index/column number).}

\item{panel}{The panel to visualize}
}
\value{
A \code{\link[patchwork:patchwork-package]{patchwork::patchwork()}} panel of graphs representing a rug plot.
}
\description{
Provide slider-style readout of proteomics panel
components.
}
\examples{
\dontrun{
plot_rug(lscc, pnum = 1, panel = "Tissue QC")
}
}
