% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyPivot.R
\name{surveyPivot}
\alias{surveyPivot}
\title{surveyPivot}
\usage{
surveyPivot(
  data,
  group_column = 1,
  response_column = 2,
  sep = ",",
  NAtoZero = TRUE,
  totals = TRUE,
  percent = TRUE
)
}
\arguments{
\item{data}{Tidy dataframe (ideally of two columns) containing group variable (e.g., ethnicity) and response variable where each row is a unique respondent.}

\item{group_column}{Column of the dataframe where the group variable resides. Defaults to 1.}

\item{response_column}{Column of the dataframe where the response variable resides. Defaults to 2.}

\item{sep}{Do you wish to save the plot output to the session directory? Default is FALSE}

\item{NAtoZero}{Do you want NA values in the pivot table to be replaced with 0? Defaults to TRUE}

\item{totals}{Do you want a column containing row sums across all groups? Defaults to TRUE.}

\item{percent}{Do you want to view the results as group response rates to each selection? Defaults TRUE}
}
\description{
Survey Pivot Function
}
\details{
Creates pivot table from survey data containing irregular multiple responses. Input is clean dataframe containing a group variable (e.g., race/ethnicity) and a response variable (selected responses to a survey question).
}
\examples{
# Basic example with all defaults unchanged
\dontrun{
survey_data <- data.frame(ethnicity = c('black','black','white','white','white'),
experience = c('I did','did not','did not','I did','did not'))
surveyPivot(survey_data)

# Adjusting defaults
surveyPivot(survey_data, group_column = 2, response_column = 1,
NAtoZero = FALSE, totals = FALSE, percent = FALSE)
}
}
\keyword{engagement}
